+++ test/prepostscripts/preLoginExample.js
@@ -10,7 +10,7 @@ module.exports = {
       // Go to Wikipedias login URL
       return driver.get('https://en.wikipedia.org/w/index.php?title=Special:UserLogin&returnto=Main+Page')
         .then(() => {
-          // You need to find the form, the login input fiels and the
+          // You need to find the form, the login input fields and the
           // password field. Just add you name and password and submit the form
           // For more docs, checkout the NodeJS Selenium version
           // http://selenium.googlecode.com/git/docs/api/javascript/index.html
@@ -20,12 +20,13 @@ module.exports = {
           // before you start, make your username and password
           var userName = 'YOUR_USERNAME_HERE';
           var password = 'YOUR_PASSWORD_HERE';
-          var loginForm = driver.findElement(webdriver.By.tagName('form'));
+          var loginForm = driver.findElement(webdriver.By.css('form'));
           var loginInput = driver.findElement(webdriver.By.id('wpName1'));
           loginInput.sendKeys(userName);
           var passwordInput = driver.findElement(webdriver.By.id('wpPassword1'));
           passwordInput.sendKeys(password);
-          loginForm.click();
+          return loginForm.click();
+          // this example skips waiting for the next page and validating that the login was successful.
         });
     })
   }